backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
langchain-anthropic>=0.3.0
httpx>=0.28.0
python-multipart>=0.0.20
